+++ controllers/users.js
@@ -1,15 +1,18 @@
 import Users from '../models/users.js';
+import CustomError from '../utils/CustomError.js';
 
 export const getUserData = async (req, res, next) => {
   try {
     const {id} = req.params;
     if (!id) {
-      return res.status(404).json({message: 'User not found'});
+      throw new CustomError('User ID is required', 400);
     }
+
     const user = await Users.findById(id).select('-password');
     if (!user) {
-      return res.status(404).json({message: 'User not found'});
+      throw new CustomError('User not found', 404);
     }
+
     res.json(user);
   } catch (error) {
     next(error);
@@ -20,24 +23,21 @@ export const updateUser = async (req, res, next) => {
   try {
     const {id} = req.params;
     const updatedData = req.body;
+
     if (!id) {
-      return res.status(404).json({message: 'User not found'});
+      throw new CustomError('User ID is required', 400);
     }
-    if (updatedData.password) delete updatedData.password;
 
-    // let user = await Users.findById(id);
-    // if (!user) {
-    //   return res.status(404).json({message: 'User not found'});
-    // }
+    if (updatedData.password) {
+      delete updatedData.password;
+    }
 
-    // // update user fields
-    // user = updatedData ? structuredClone(updatedData) : {};
-    // console.log(user);
-    // await user.save(); // Save updated user to the database
+    const updatedUser = await Users.findByIdAndUpdate(id, updatedData, {new: true});
+    if (!updatedUser) {
+      throw new CustomError('User not found', 404);
+    }
 
-    const updateUser = await Users.findByIdAndUpdate(id, updatedData, {new: true});
-    console.log(updateUser);
-    res.json({message: 'User updated successfully', user: updateUser});
+    res.json({message: 'User updated successfully', user: updatedUser});
   } catch (error) {
     next(error);
   }
@@ -47,8 +47,15 @@ export const deleteUser = async (req, res, next) => {
   try {
     const {id} = req.params;
 
+    if (!id) {
+      throw new CustomError('User ID is required', 400);
+    }
+
     const deletedUser = await Users.findByIdAndDelete(id);
-    if (!deletedUser) return res.status(404).json({message: 'User not found'});
+    if (!deletedUser) {
+      throw new CustomError('User not found', 404);
+    }
+
     res.json({message: 'Account deleted successfully'});
   } catch (error) {
     next(error);
